fix: remove WatermelonDB internals from batch pullChanges
We let `_changed` and `_status` through to what we stored in the cloud for batched writes, which WatermelonDB doesn't like when you import those records. So we now remove those fields even when pulling down from the cloud, and also remove them from batched writes (which was the bug in the first place).